move readTsdb to parquet storage sub-package; rename method & vars for clarity
