+++ pkg/storage/parquet/cache.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
 
-package storage
+package parquet
 
 import (
 	lru "github.com/hashicorp/golang-lru/v2"
