+++ pkg/storage/parquet/row.go
@@ -3,7 +3,7 @@
 // Provenance-includes-license: Apache-2.0
 // Provenance-includes-copyright: The Cortex Authors.
 
-package storage
+package parquet
 
 const (
 	ColHash                        = "s_hash"
